ci: split global worktree handlers
